markerget functionality
Added markerget function, which parses through an input file name given a marker, and then takes the value for that number. If you follow a file naming system with markers, this allows the time series program (for instance) to change data like number of APs instead of making you hard-code each time.